requirements.txt
nltk
discord.py
markovify
colorama
discord-py-interactions==3.0.2